refactor(tui): extract chat stream and tool message rendering component
